perf(inter): add HashTable for inter(and ghost) atoms to reduce the time complexity of inter atoms searching.
In each time step, we build a hash table for all inter atoms and inter ghost atoms. When calculating
EAM in atom::computeEam, we can compute Rho and Force of inter atoms to inter atoms in O(M) time
(where M is the count of inter atoms), while in the old method the time complexity is O(M^2).

fixed #9