fix(ws): fix bug of converting atom postion to lattice index.
In fact, the lattice index(x,y,z) is different for even and odd lattice x index in BCC lattice.
In this change, we add judgment for even and odd lattice x index.
But after this change, the system sum force are not allways equals to zero(maybe another bug).