feat(packer): send atom id to neighbor processes as id of ghost atoms if it's under development mode
If it is under development mode, the macro `DEV_MD_COMM_INC_ATOM_ID` can be used, then it will pack
atom ids and send to neighbor processes when communicating ghost atoms, and for receiving process,
it will unpack atom ids and set them as id of ghost atoms. If it is not development mode, atom ids
will not be packed/unpacked when communicating ghost atoms (id of all ghost atoms will be set to 0).

this commit add support for communicating id of both lattice atoms and inter-atoms under development
mode.