feat(output): add ability to use `preset` dump config to output atoms data
In this commit, we add ability to set `presets` for atom dump, which is similar as a template. In
simulation stage, we can use the preset via `use` keword and set `every_steps` for the dump.

BREAKING CHANGE: In this commit, we removed the global dump config. Instead, we use "dump preset" to
dump system atoms to file